filesystem/iso9660: rework & fix SU area iteration
Previously we wouldn't have any real indication as to whether
the current iteration context is holding a reference to the
ca_cache, which would lead to relying on very unreliable signs
such as current offset and location within a continuation area.
Because of these vague reference owning semantics, the
current implementation had a bug where a reference would
either get leaked or double released depending on the
situation.

Fix this by introducing a 'has_ref' member to the susp iteration
context, and make it active even for cases when the context is
currently not in a continuation area to better verify the
invariants currently present in the code.