tests: add a base loader test suite & a test kernel
This introduces a basic set of tests for the loader driven by pytest that
includes booting a test runner kernel (also introduced with this commit),
which verifies the entire boot context & memory map for sanity and validity
and reports the test status accordingly afterwards.

Current tests include fat12/16/32 & iso9660 cases that run both UEFI & BIOS
loader configurations with more tests coming later down the line. The test
runner is designed to be able to skip/select tests based on what is available
in the current environment. For example, it will simply skip all of the tests
that utilize UEFI if it's unable to find a UEFI firmware file.

The kernel utilized by the test cases is an ultra protocol compatible kernel
that can be built in i686, amd64_lower_half, as well as amd64_higher_half
modes. It also supports basic 0xE9 & linear framebuffer logging, which makes
it viable to run it on real hardware as well. The kernel can automatically
detect whether it is running inside a hypervisor and alter its behavior
accordingly. It also supports a 'no-shutdown' command line parameter for
convenience.