filesystem/iso9660: introduce proper dir-record ref owning semantics
Previously we would release the reference before returning to the
caller, which meant it would have to use a dangling reference to
block_cache in order to parse the directory record.

Fix this issue by introducing a similar concept to what was done for
susp_iteration_ctx. There is now a 'has_ref' member as part of the
directory iteration context, which is used to enforce proper reference
owning semantics.

Also add calls to acquire/release ref where appropriate, as well as
additional invariant validation in dir_iter_ctx_acquire_ref.