uefi: disable stack arg probing
Stack arg probing is intended to be used in a hosted ring3 environment
(aka a normal Windows userland application) when there's a kernel
managing stack commit & guard pages.

Function prologues emit calls to __chkstk and __chkstk_ms when this
option is enabled.

The default implementation of the aforementioned functions in
libgcc/cygwin.S has a comment explaining stack probing further:

   Function prologue calls __chkstk to probe the stack when allocating more
   than CHECK_STACK_LIMIT bytes in one go.  Touching the stack at 4K
   increments is necessary to ensure that the guard pages used
   by the OS virtual memory manger are allocated in correct sequence.

UEFI environments shouldn't need this mechanism as stack pages are always
mapped in.