Modularize file system registration & detection code
Make it more dynamic by introducing DECLARE_FILESYSTEM, and a
struct filesystem_type that provides all the necessary information
to detect & initialize a file system.

This removes the annoying boilerplate where every file system had to
make its own header with a detect function that had to be manually
called in the fs detection code. Now all that a file system has to do
is to make its type & declare it, which will be picked up automatically
by the fs probe code.