tests/disk_image: preserve GPT backup headers when formatting an image
Add an extra megabyte at the end to make sure GPT headers are intact (this
might be a bit overkill, but at least it keeps the image 1M aligned).

This also adds a partition length parameter to image_partition to make sure the
extra padding at the end stays empty and the partition only occupies space that
we actually need. Previously we would make it as big as possible.