loader: configure link options per platform
Make each platform set its own link flags so that specific
workarounds only have to be set in that platform's
cmakelists, and so that we don't have to assume as much
internal details about how a specific platform executable
gets linked.

As an added bonus this replaces some UEFI-specific link
flags to use their full verbose versions for better readibility.