Add a new dummy toolchain file for IDE work
The toolchain_dummy.cmake file is used for configuring
the project when running under an IDE, which doesn't
necessarily understand all the compile options, and
doesn't have to be able to actually build the project,
since it is only used for the sake of syntax highlighting
and code completions.

This allows it to still pick up all the necessary files
without failing midway because of an unknown compile
flag or option.

As an added bonus, this automatically detects CLion
and makes it use the dummy toolchain file.