ultra: factor out module data allocation & use new allocation API
This code was previously duplicated in multiple branches of the module
loading code in module_load() with only minor differences in how the
memory was being zeroed.

Make a separate helper module_data_alloc(), which takes take of memory
allocation as well as zero-filling it.

As an added bonus this also slightly reworks module_get_load_address to
honor cases where configuration explicitly sets load address to zero.