config: make config freeable & highlight data ownership
This commit introduces the ability to free a configuration file
including its source.

Previously, the source wasn't stored directly in the config structure
but referenced by it nonetheless with an implicit assumption that it
will remain alive at least for the duration of the config struct
lifetime.

Make this more explicit by introducing a new struct called config_source
and storing it directly in the config structure.

Also introduce a new 'cfg_release' helper, which takes care of freeing
both the underlying entries buffer as well as the source itself.