loader: introduce cleanup handlers & 'services_cleanup'
Cleanup handlers are registered by services or helpers that require
dynamic resource allocations in order to work, and that do not affect
memory management and/or handover services.

The new service function runs all registered clean up handlers with the
goal to minimize the amount of loader-reclaim memory that's passed on to
the kernel in the memory map.

The caller must assume that no services are usable after this call
besides the memory management & handover.