ultra: fix command line use-after-free
We were referencing a string from inside the configuration file in
build_attribute_array(), which is called after we release the config
thus causing a use-after-free.

Fix this by introducing a fixed-capacity buffer on the stack that we use
as the internal storage for the command line (if any). While at it also
make a separate function for retrieving the command line itself so that
it's easier to follow the configuration->internal buffer dance that we
have to do.