memory_services: introduce mm_foreach_entry & remove old API
This commit adds new API that allows for memory map iteration with a
callback without having to run services_release_resources. This also
removes the now superseded 'ms_get_highest_map_address' API, which can
be implemented entirely in terms of 'mm_foreach_entry' if needed.

One of the intended use cases of 'mm_foreach_entry' is for protocols
that support virtual memory and need to be able to decide what physical
ranges of memory are to be mapped-in.

On top of that, make the ultra protocol implementation use this new API,
and adjust its behavior so that it only maps whatever is present in the
memory map as opposed to everything up to the last byte. This conforms
to the behavior change in UltraProtocol commit 39129a35339 ("Change vm-mapping guarantees to be more performance friendly")