virtual_memory: rewrite to be less messy and remove helpers
Rewrite in a similar fashion as it was done for allocator & elf
utilities. All the parameters are now passed via a generic
page_mapping_spec, which is easy to extend and fill.

All of the previosuly existing helpers are no longer needed as they can
all be flexibly emulated via the new mapping spec structure. Since page
mapping is not something that is used in a lot of places around the
codebase we don't need to have any narrower scope convenience helpers as
we did in the allocation code.