virtual_memory: map all pages in bulk
This greately improves page mapping performance as we no longer have to
walk the entire multi-level table each time, instead we do this per N
entries where N is the number of entries in a table.

E.g previously if we wanted to map 1024 pages in an AMD64 4 level page
table we would walk the entire table 1024 times, whereas now we only
have to walk it 2 times since we map 512 pages at a time.