elf: don't require the entire binary to be preloaded in memory
Instead, load only what's needed and use a separate block_cache for
iterating meta header information and/or program headers.

Add a new elf_io structure to handle this, as well as a few helpers that
are used to initialize it.

Rewrite parts of the code that relied on the entire binary to be
preloaded to instead use the new elf_io_take_ref/elf_io_unref helpers.