uefi/cmake: only allow general purpose registers
UEFI makes no particular guarantees about SSE being enabled on 64-bit
platforms, even if a CPU has it. The only place where this is
technically useful is copying & zeroing memory and we already use the
CopyMem & SetMem functions provided by the UEFI runtime, which should be
implemented with SSE if it is supported anyway.

This also makes the u-boot UEFI implementation bootable.