Resturcture the project to arch-specific folders
So that no x86-specific stuff leaks out into generic code and so that
we have an easier time extending support for other architectures. This
also cleans up the project structure a little bit so that all e.g all
bios stuff is inside the bios sudirectory as opposed to being in the
project root for some of it.