virtual_memory: add a new entry_address_mask field to page_table
This mask is used to extract the actual physical address from a page
table entry. Previously we would rely on all of the top bits being clear
so we didn't need it. This is a fragile assumption that also breaks on
other architectures that use top bits more frequently, unlike x86, which
mostly uses it for NX (that we ignore entirely).