common: introduce rw_helpers
These are basic read/write helpers that take in a void pointer and
access it with the specified width either by reading or writing.

Similar helpers have been duplicated in the loader in a few places
including virtual_memory for read/write_slot and the UEFI relocator
utility, which used them for updating a user value with the relocated
pointer address. These are going to be refactored to use the new
rw_helpers in the following commits.