handover: store page_table directly in the handover info
This simplifies a lot of things like having to pass the page table
around and then losing all of the information about it when we're left
with just the bare pt_root physical address. This doesn't work for all
architectures that might need additional information about the page
table after we actually construct it.