boot_protocol/ultra: fill out the kernel info attr paths correctly
Previously we would fill them out incorrect in case the path was
specified as ORIGIN, because the fields of the path are not filled fully
in that case. We were also forgetting to set the disk_index field
completly.

Fix this by treating path type ORIGIN as a special case.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>